requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-environ==0.11.2
django-tinymce==4.1.0
easy-thumbnails==2.8.5
idna==3.7
pillow==10.3.0
psycopg2==2.9.9
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
gunicorn